+++ tests/testthat/test_gtdd.R
@@ -3,12 +3,12 @@ library(GetTDData)
 
 dl.folder <- 'TD Data'
 
-my.flag <- download.TD.data(asset.codes = 'LTN_2015', dl.folder = dl.folder)
+my.flag <- download.TD.data(asset.codes = 'LTN', dl.folder = dl.folder, n.dl = 1)
 test_that(desc = 'Test of download function',{
           expect_equal(my.flag , TRUE) } )
 
 
-returned.rows <- nrow(read.TD.files(maturity = '010116', dl.folder = dl.folder ))
+returned.rows <- nrow(read.TD.files(maturity = '010117', dl.folder = dl.folder ))
 
 test_that(desc = 'Test of read function',{
   expect_equal(returned.rows>0, TRUE)